+++ lib/modules/i18n/i18n_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: i18n
-// Updated at: 2026-05-10 09:18:49
-// Commit: feat(i18n): add bi-directional event bus handler for i18n
+// Updated at: 2026-05-10 10:43:50
+// Commit: feat(i18n): implement shimmer loading skeleton in i18n
 
 class i18nComponentState {
-  final String commitRef = '987';
-  final String lastMessage = 'feat(i18n): add bi-directional event bus handler for i18n';
-  final DateTime timestamp = DateTime.parse('2026-05-10T09:18:49+05:00');
+  final String commitRef = '990';
+  final String lastMessage = 'feat(i18n): implement shimmer loading skeleton in i18n';
+  final DateTime timestamp = DateTime.parse('2026-05-10T10:43:50+05:00');
 }